Add files via upload
Every company wants to increase its revenue and profitability. To do that, while they acquire new customers, they also want to make sure that the existing ones stay with them for a long term. Also, its strategically important to know beforehand whether a set of customers are planning to stop using their services (especially recurring ones like internet, cable, phone etc.). To do that, every company or business creates and tracks customer metrics which are then used to predict their likelihood of churn.

Customer Churn for a company occurs when a customer decides to stop using the services of that company. In this project, we will be using the customer data of a telecom sector company based in the US to predict the probability of churn for each of the customer. We will look at the standard practices that are followed in the industry to solve these problems and also go beyond just those techniques. We have chosen the telecom company data for the churn problem as it is a major area of concern for companies in that sector.

Once we have built a model, the churn model output can also be used as a warning indicator that some customers are likely to churn. The key drivers that are making the customer more likely to churn can be alleviated and ensure that the customers are actually retained.


Understand the Problem Statement
Perform basic EDA to get familiarize with the data
Take care of any missing values or datatype issues in the data
Perform Univariate analysis for both numeric and categorical variables
Perform Bi-variate analysis to identify redundant variables
Plot Probability Distribution Function (PDF) of each predictor the target variable
Create new features that might add value to the model
Define a function for each set of code that might need to be repeated again
Prepare the data for modelling
Create models using multiple algorithms
Perform Hyper-parameter tuning to get the best parameters
Compare the performance of different models using multiple metrics
Apply feature reduction methods to get a light model while maintaining performance
Check whether the top model features are as per your business understanding/intuition
Save the model and creating API to implement the model in production