Price Optimization
Understanding the retail price optimization machine learning problem
In depth understanding of price elasticity
Settings for Jupyter notebook to enablefaster coding
Importing Libraries
Importing datasets and initial understanding with the help of visualizations
Understanding business context with the help of data
Combining data
Making inferences from plots
Learning to segregate data based on analysis
Implementing model to identify price elasticity of items
Creating generic code to identify price elasticity of all items
Understand criteria to select model
Walkthrough of price optimization for one product and visualizing the outputs
Generic code for price optimization for all products


Pricing a product is a crucial aspect in any business. A lot of thought process is put into it. There are different strategies to price different kinds of products. There are products whose sales are quite sensitive to their prices and as such a small change in their price can lead to noticeable change in their sales. While there are products whose sales are not much affected by their price - these tend to be either luxury items or necessities (like certain medicines). This machine learning retail price optimization project will focus on the former type of products.

Price elasticity of demand (Epd), or elasticity, is the degree to which the effective desire for something changes as its price changes. In general, people desire things less as those things become more expensive. However, for some products, the customers desire could drop sharply even with a little price increase, and for other products, it could stay almost the same even with a big price increase. Economists use the term elasticity to denote this sensitivity to price increases. More precisely, price elasticity gives the percentage change in quantity demanded when there is a one percent increase in price, holding everything else constant.

In this machine learning pricing optimization case study, we will take the data of a cafe and based on their past sales, identify the optimal prices for their items based on the price elasticity of the items. For each item, first the price elasticity will be calculated and then the optimal price will be figured. While this is taking a particular cafe data, this work can be extended to price any product.